Implemented new silver short sword artifact: Insanity's Edge.
It causes conflict (w/o hunger), and it stuns its targets, but it causes
the wielder to hallucinate - unless the wielder is blind.

This has been added to the _end_ of the artifacts list; and so hopefuly
it will not break save / bones compatibility. (I don't know for sure,
but I think it will work ok.)