Adjusted the stun effect of Helm of Anti-magic.
The effect is now handled by a new function: magic_resisted_effect; and
the effect now depends on the particular type of magic which was
resisted. For example, resisting a death effect will stun more often
than reflecting magic missile damage will.

Also, the stun will no longer override pre-existing stun effects, and it
will not be applied when magic resistance is not need to resist the
effect (eg. resisting paralyse when the user also has 'free action'.)