Rewrote the scoring system.
The new scoring system will generally produce significantly lower scores
than before. Unfortunately, this means that any existing high-score
tables are essentially invalid. Old high-score files will still "work",
but I recommend that they be deleted so that the newer score can be
recorded.

Also, save games loaded from previous versions of K-Mod may end up with
larger than intended scores - so those scores should be manually removed
from the new high score table as well.

--

The new scoring system gives points for all vanquished monsters, not
just those killed directly by the player.

Points for vanquished creatures are calculated using the base stats of a
monster rather than the stats of the specific incarnations that died (so
equipment and level changes will not affect the final score).

The previous scoring system heavily favoured characters who killed
enemies directly, rather than with pets or 'conflict'. But more
importantly, the previous system was unfair in that if a pet happened to
get the first kill of a particular species, the player was effectively
robbed of the full points of that species (because the points depends on
how many of the species have died, and the first one is worth the most
points by far). The new scoring system removes these issues.

Also, the points per kill is roughly 1/4 of what it was previously -
this means that treasure now plays a much greater role in the final
score.

--

Dump files now include the version number.